test(cli): split tunnel credentials suite
